More helpers and updates on existing tests
