+++ e2e/tests/tickets.e2e.ts
@@ -1,13 +1,12 @@
-import {device, by, element} from 'detox';
-import {findTextViewElementDisplayText, goBack, goToTab} from '../utils/commonHelpers';
+import {device} from 'detox';
+import {goToTab} from '../utils/commonHelpers';
+import {tapById, tapByText} from '../utils/interactionHelpers';
 import {
   expectToBeVisibleById,
-  expectNotToBeVisibleById,
   expectToBeVisibleByText,
-  tapById,
-  tapByText,
   expectTextById,
-} from '../utils/interactionHelpers';
+  expectToBeVisibleByPartOfText,
+} from '../utils/expectHelpers';
 import {skipOnboarding} from '../utils/onboarding';
 import setLocation from '../utils';
 
@@ -22,7 +21,7 @@ describe('Tickets', () => {
         locale: 'US',
       },
     });
-    //await setLocation(62.4305, 9.3951);
+    await setLocation(62.4305, 9.3951);
     await skipOnboarding();
 
     // Accept ticket limitations
@@ -51,8 +50,8 @@ describe('Tickets', () => {
     await expectToBeVisibleByText('Single ticket bus/tram');
     await expectToBeVisibleByText('1 Adult');
     await expectToBeVisibleByText('Starting now');
-    //Sometimes C3 comes up in the tests
-    //await expectToBeVisibleByText('Travel in 1 zone (A)');
+    //Zone is either A or C3 during the tests
+    await expectToBeVisibleByPartOfText('Travel in 1 zone');
     await expectTextById('offerTotalPriceText', `Total: ${ticketPrice}`);
 
     await tapById('goToPaymentButton');
@@ -61,7 +60,8 @@ describe('Tickets', () => {
     await expectToBeVisibleByText('1 Adult');
     await expectToBeVisibleByText(ticketPrice);
     await expectToBeVisibleByText('Single ticket bus/tram');
-    await expectToBeVisibleByText('Valid in zone A');
+    //Zone is either A or C3 during the tests
+    await expectToBeVisibleByPartOfText('Valid in zone');
     await expectToBeVisibleByText('Starting now');
 
     // Check the payment cards
@@ -107,7 +107,7 @@ describe('Tickets', () => {
     }
   });
 
-  it('should be able to buy a period ticket', async () => {
+  xit('should be able to buy a period ticket', async () => {
     await goToTab('tickets');
 
     await expectToBeVisibleByText('Buy');
